Ver 3.1.4
Jzip 3.1.4 改进细节和预防潜在问题。

1.添加了 Jzip 便携版更新提醒，不在默认路径更新 Jzip 前清空文件夹作出提示。
2. Jzip 命令行帮助改动为不弹出窗口列出。
3.修复了打开 .msi 文件时查看器为空的问题。

请注意 Jzip 便携版更新前会清空 Jzip 所在文件夹。确保路径不含个人文件。

+++ App/Parts/VerControl.cmd
@@ -21,42 +21,52 @@ set "if.error.2=|| ( call :MsgBox "下载的文件不存在，请重新尝试。" "%if.error.lm
 set "if.error.3=|| ( call :MsgBox "更新包出现错误，请重新尝试。" "%if.error.lm%" & goto :EOF )"
 set "if.error.4=|| ( call :MsgBox "缺失 Bitsadmin 组件，在早期版本 Windows 中不存在。" "%if.error.lm%" & goto :EOF )"
 
-:: 安装时，配置路径和窗口
+:: 配置路径和窗口
+set "dir.jzip.default=%appdata%\JFsoft\JZip\App"
+
 if "%1"=="Install" (
-	set "dir.jzip=%appdata%\JFsoft\JZip\App"
+	set "dir.jzip=%dir.jzip.default%"
 	set "dir.jzip.temp=%temp%\JFsoft.JZip"
 	>nul 2>nul md !dir.jzip.temp!
 	mode 80, 25
 	color f0
 )
-	
-:: 检测 Bits 组件
-bitsadmin /? >nul 2>nul %if.error.4%
 
-:: 若 Bits 服务被禁用，询问开启
-sc qc bits | findstr /i "DISABLED" >nul && (
-	if "%1"=="Install" call :MsgBox-s key "Jzip 安装过程需要 Bits 服务。" " 您是否允许 Jzip 启用服务？"
-	if "%1"=="Upgrade" call :MsgBox-s key "Jzip 更新过程需要 Bits 服务。" "您是否允许 Jzip 启用服务？"
-	
-	if "!key!"=="1" (
-		:: 启用 Bits 服务
-		net session >nul 2>nul && (sc config bits start= demand >nul)
-		net session >nul 2>nul || (
-			mshta vbscript:CreateObject^("Shell.Application"^).ShellExecute^("cmd.exe","/c sc config bits start= demand >nul","","runas",1^)^(window.close^)
+:: Jzip 便携版判断
+if /i "%dir.jzip%"=="%dir.jzip.default%" (set "jzip.Portable=") else (set "jzip.Portable=1")
+
+:: Mshta 可用性判断
+mshta /? || ( echo.Mshta 不可用，无法安装 JZip。 & echo.%if.error.lm% & pause >nul )
+
+:: 检测 Bits 组件
+for %%i in (Install Upgrade) do if "%1"=="%%i" (
+	bitsadmin /? >nul 2>nul %if.error.4%
+
+	:: 若 Bits 服务被禁用，询问开启
+	sc qc bits | findstr /i "DISABLED" >nul && (
+		if "%1"=="Install" call :MsgBox-s key "Jzip 安装过程需要 Bits 服务。" " 您是否允许 Jzip 启用服务？"
+		if "%1"=="Upgrade" call :MsgBox-s key "Jzip 更新过程需要 Bits 服务。" "您是否允许 Jzip 启用服务？"
+
+		if "!key!"=="1" (
+			:: 启用 Bits 服务
+			net session >nul 2>nul && (sc config bits start= demand >nul)
+			net session >nul 2>nul || (
+				mshta vbscript:CreateObject^("Shell.Application"^).ShellExecute^("cmd.exe","/c sc config bits start= demand >nul","","runas",1^)^(window.close^)
+			)
+			ping localhost -n 2 >nul
+		) else (
+			call :MsgBox "抱歉，无法安装 JZip。" "%if.error.lm%"
+			goto :EOF
 		)
-		ping localhost -n 2 >nul
-	) else (
-		call :MsgBox "抱歉，无法安装 JZip。" "%if.error.lm%"
-		goto :EOF
 	)
 )
 
 :: 获取 Github 上的 JZip 安装信息
-for %%a in (Install Upgrade) do if "%1"=="%%a" (
-	dir "%dir.jzip.temp%\ver.ini" /a:-d /b >nul 2>nul && del /q /f /s "%dir.jzip.temp%\ver.ini" >nul 2>nul
+for %%i in (Install Upgrade) do if "%1"=="%%i" (
+	>nul 2>nul ( dir "%dir.jzip.temp%\ver.ini" /a:-d /b && del /q /f /s "%dir.jzip.temp%\ver.ini" )
 	bitsadmin /transfer !random! /download /priority foreground https://raw.githubusercontent.com/Dennishaha/JZip/master/Server/ver.ini "%dir.jzip.temp%\ver.ini" %if.error.1%
 	cls
-	dir "%dir.jzip.temp%\ver.ini" /a:-d /b >nul 2>nul %if.error.2%
+	>nul 2>nul dir "%dir.jzip.temp%\ver.ini" /a:-d /b %if.error.2%
 
 	for /f "eol=[ usebackq tokens=1,2* delims==" %%a in (`type "%dir.jzip.temp%\ver.ini"`) do set "%%a=%%b"
 	set "jzip.newver.page=%dir.jzip.temp%\full.!jzip.newver!.exe"
@@ -85,40 +95,50 @@ for %%a in (Install Upgrade) do if "%1"=="%%a" if /i not "%jzip.ver%"=="%jzip.ne
 ::UI--------------------------------------------------
 
 :: 弹出选择
-if "%1"=="Install" call :MsgBox-s key "现在可以安装 Jzip %jzip.newver%。"
-if "%1"=="UnInstall" call :MsgBox-s key "确实要卸载 JZip 吗？"
+if "%1"=="Install" call :MsgBox-s key "现在可以安装 Jzip %jzip.newver%。" " " "安装将自动开始。"
+if "%1"=="UnInstall" call :MsgBox-s key "确实要解除安装 JZip 吗？"
 if "%1"=="Upgrade" if /i "%jzip.ver%"=="%jzip.newver%" call :MsgBox "JZip %jzip.ver% 是最新的。" & goto :EOF
-if "%1"=="Upgrade" if /i not "%jzip.ver%"=="%jzip.newver%" call :MsgBox-s key "现在可以获取新版本 JZip %jzip.newver%。"
+if "%1"=="Upgrade" if /i not "%jzip.ver%"=="%jzip.newver%" call :MsgBox-s key "现在可以获取新版本 JZip %jzip.newver%。" " " "安装将自动开始。"
 if not "%key%"=="1" goto :EOF
 
 :: 获取 JZip 安装包
 for %%a in (Install Upgrade) do  if "%1"=="%%a" (
-	dir "%jzip.newver.page%" /a:-d /b >nul 2>nul && del /q /f /s "%jzip.newver.page%" >nul 2>nul
+	>nul 2>nul ( dir "%jzip.newver.page%" /a:-d /b && del /q /f /s "%jzip.newver.page%" )
 	bitsadmin /transfer %random% /download /priority foreground %jzip.newver.url% "%jzip.newver.page%" %if.error.1%
 	cls
-	dir "%jzip.newver.page%" /a:-d /b >nul 2>nul %if.error.2%
+	>nul 2>nul dir "%jzip.newver.page%" /a:-d /b %if.error.2%
 	"%jzip.newver.page%" t | findstr "^Everything is Ok" >nul 2>nul %if.error.3%
 )
 
 :: 解除安装
-for %%a in (UnInstall Upgrade) do if "%1"=="%%a" (
+for %%a in (Upgrade UnInstall) do if "%1"=="%%a" (
 	call "%dir.jzip%\Parts\Set_Lnk.cmd" -off all
 	call "%dir.jzip%\Parts\Set_Assoc.cmd" & if "!tips.FileAssoc!"=="●" call "%dir.jzip%\Parts\Set_Assoc.cmd" -off
 )
 
 :: 安装
 for %%a in (Install Upgrade) do if "%1"=="%%a" (
-	cmd /q /c "rd /q /s "%dir.jzip%" >nul 2>nul & md "%dir.jzip%" >nul 2>nul & "%jzip.newver.page%" x -o"%dir.jzip%\" & "%dir.jzip%\%jzip.newver.installer%" -install"
+
+	:: Jzip 便携版判断 
+	set "key="
+	if defined jzip.Portable call :MsgBox-s key "您正使用 JZip 便携版，更新前将清空 JZip 所在文件夹。" " " "%dir.jzip%" "请确保路径不含个人文件。" " " "确定吗？"
+
+	if not "!key!"=="2" cmd /q /c "rd /q /s "%dir.jzip%" >nul 2>nul & md "%dir.jzip%" >nul 2>nul & "%jzip.newver.page%" x -o"%dir.jzip%\" & "%dir.jzip%\%jzip.newver.installer%" -install"
 	exit
 )
 
-:: 删除 JZip 目录
+:: 删除 JZip 注册表项和目录
 if "%1"=="UnInstall" (
 	>nul (
 		reg delete "HKCU\Console\JFsoft.Jzip" /f
 		reg delete "HKCU\Software\JFsoft.Jzip" /f
 	)
-	cmd /q /c "rd /q /s "%dir.jzip%"  >nul 2>nul"
+
+	:: Jzip 便携版判断
+	set "key="
+	if defined jzip.Portable call :MsgBox-s key "已完成 Jzip 便携版解除安装。" " " "移除 JZip 所在路径吗？" " " "%dir.jzip%" "请确保路径不含个人文件。" " " "确定吗？"
+	
+	if not "!key!"=="2" start "" /min cmd /q /c ">nul rd /q /s "%dir.jzip%""
 )
 goto :EOF
